refactor(ep): move the ORT API version negotiation into ort-bridge
The negotiation code sat in morphizen-ort-api-ext, whose purpose is the
MorphizenOrtApiExt table that passes and graph components consume. Version
negotiation is instead a plugin-entry concern: its only callers are the four
ort-bridge sources that build ABI structs, and it must run from
CreateEpFactories. Keeping it in the shared library also put the header on a
PUBLIC include path for every component that links it.

Upstream keeps the equivalent code in src/shared/common because four backend
DLLs share it; hipgpu.dll has a single plugin entry, so ort-bridge/src is the
narrower home.

No behaviour change: rebuilt and re-verified against ORT 1.27.0, 1.24.2 and
1.23.2 with the same results.
