+++ morphizen/ort-bridge/src/morphizen-ep.cpp
@@ -5,13 +5,13 @@
 #include "./morphizen-ep.hpp"
 #include "./ir-converter.hpp"
 #include "./morphizen-deps.hpp"
+#include "./ort-api-version.hpp"
 #include "./ort-graph-wrapper.hpp"
 #include "glog/logging.h"
 #include "morphizen-utils/morphizen-utils.hpp"
 #include "morphizen/morphizen-ort-api-ext.hpp"
 #include "morphizen/morphizen.hpp"
 #include "morphizen/onnxruntime_morphizen_ep.hpp"
-#include "morphizen/ort-api-version.hpp"
 #include <google/protobuf/util/json_util.h>
 #include <set>
